chat-gateway/requirements.txt
openai>=1.0.0
Flask
flask-cors
firebase-admin
python-dotenv
requests
